TFM/requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-google==10.21.1
apache-airflow-providers-http==4.7.0
beautifulsoup4
requests
fastapi 
uvicorn 
pandas